chore: create execution_resources file to split the starknet api transaction file
