+++ crates/papyrus_test_utils/src/lib.rs
@@ -85,6 +85,7 @@ use starknet_api::deprecated_contract_class::{
     StructType,
     TypedParameter,
 };
+use starknet_api::execution_resources::{Builtin, ExecutionResources, GasVector};
 use starknet_api::felt;
 use starknet_api::hash::{PoseidonHash, StarkHash};
 use starknet_api::state::{
@@ -98,7 +99,6 @@ use starknet_api::state::{
 };
 use starknet_api::transaction::{
     AccountDeploymentData,
-    Builtin,
     Calldata,
     ContractAddressSalt,
     DeclareTransaction,
@@ -117,9 +117,7 @@ use starknet_api::transaction::{
     EventData,
     EventIndexInTransactionOutput,
     EventKey,
-    ExecutionResources,
     Fee,
-    GasVector,
     InvokeTransaction,
     InvokeTransactionOutput,
     InvokeTransactionV0,
