+++ crates/starknet_api/src/lib.rs
@@ -9,6 +9,7 @@ pub mod crypto;
 pub mod data_availability;
 pub mod deprecated_contract_class;
 pub mod executable_transaction;
+pub mod execution_resources;
 pub mod hash;
 pub mod rpc_transaction;
 pub mod serde_utils;
