+++ crates/starknet_client/src/reader/objects/transaction.rs
@@ -14,6 +14,7 @@ use starknet_api::core::{
     EthAddress,
     Nonce,
 };
+use starknet_api::execution_resources::GasVector;
 use starknet_api::hash::StarkHash;
 use starknet_api::transaction::{
     AccountDeploymentData,
@@ -24,7 +25,6 @@ use starknet_api::transaction::{
     DeployTransactionOutput,
     Event,
     Fee,
-    GasVector,
     InvokeTransactionOutput,
     L1HandlerTransactionOutput,
     L1ToL2Payload,
@@ -664,7 +664,7 @@ pub enum Builtin {
     RangeCheck96,
 }
 
-impl From<ExecutionResources> for starknet_api::transaction::ExecutionResources {
+impl From<ExecutionResources> for starknet_api::execution_resources::ExecutionResources {
     fn from(execution_resources: ExecutionResources) -> Self {
         let da_gas_consumed = execution_resources.data_availability.unwrap_or_default();
         Self {
@@ -674,27 +674,39 @@ impl From<ExecutionResources> for starknet_api::transaction::ExecutionResources 
                 .into_iter()
                 .filter_map(|(builtin, count)| match builtin {
                     Builtin::RangeCheck => {
-                        Some((starknet_api::transaction::Builtin::RangeCheck, count))
+                        Some((starknet_api::execution_resources::Builtin::RangeCheck, count))
                     }
                     Builtin::Pedersen => {
-                        Some((starknet_api::transaction::Builtin::Pedersen, count))
+                        Some((starknet_api::execution_resources::Builtin::Pedersen, count))
                     }
                     Builtin::Poseidon => {
-                        Some((starknet_api::transaction::Builtin::Poseidon, count))
+                        Some((starknet_api::execution_resources::Builtin::Poseidon, count))
                     }
-                    Builtin::EcOp => Some((starknet_api::transaction::Builtin::EcOp, count)),
-                    Builtin::Ecdsa => Some((starknet_api::transaction::Builtin::Ecdsa, count)),
-                    Builtin::Bitwise => Some((starknet_api::transaction::Builtin::Bitwise, count)),
-                    Builtin::Keccak => Some((starknet_api::transaction::Builtin::Keccak, count)),
+                    Builtin::EcOp => {
+                        Some((starknet_api::execution_resources::Builtin::EcOp, count))
+                    }
+                    Builtin::Ecdsa => {
+                        Some((starknet_api::execution_resources::Builtin::Ecdsa, count))
+                    }
+                    Builtin::Bitwise => {
+                        Some((starknet_api::execution_resources::Builtin::Bitwise, count))
+                    }
+                    Builtin::Keccak => {
+                        Some((starknet_api::execution_resources::Builtin::Keccak, count))
+                    }
                     // output builtin should be ignored.
                     Builtin::Output => None,
                     Builtin::SegmentArena => {
-                        Some((starknet_api::transaction::Builtin::SegmentArena, count))
+                        Some((starknet_api::execution_resources::Builtin::SegmentArena, count))
                     }
-                    Builtin::AddMod => Some((starknet_api::transaction::Builtin::AddMod, count)),
-                    Builtin::MulMod => Some((starknet_api::transaction::Builtin::MulMod, count)),
+                    Builtin::AddMod => {
+                        Some((starknet_api::execution_resources::Builtin::AddMod, count))
+                    }
+                    Builtin::MulMod => {
+                        Some((starknet_api::execution_resources::Builtin::MulMod, count))
+                    }
                     Builtin::RangeCheck96 => {
-                        Some((starknet_api::transaction::Builtin::RangeCheck96, count))
+                        Some((starknet_api::execution_resources::Builtin::RangeCheck96, count))
                     }
                 })
                 .collect(),
